Software Renderer: Fix artifacts with non pixel aligned content
It was possible that we would request to flush more of the backing store
than what was actually painted, which could lead to artifacts. Now when
there is a difference caused by the pixel alighment conversions, we make
sure to mark items behind that content as dirty so that everything
necessary to render correctly is painted.

Fixes: QTBUG-133368
Pick-to: 6.10 6.8
Change-Id: I9766017eb610792ffa17b745c356b614e4e28752
Reviewed-by: Christian Strømme <christian.stromme@qt.io>