Fix QQuickIconLabel action icon color being overridden by style default
Styles need a way to set the default icon color for controls. Until
now, that was done like this:

    icon.color: "tomato"

    contentItem: IconLabel {
        icon: control.icon
    }

This breaks the use case of e.g. an Action that sets its own
icon.color, because the icon property of the control always takes
precedence (see 146bc9517c56feda4eba34282d3cc53bd47b6267).

This patch adds a defaultIconColor property to IconLabel, which allows
styles to specify a color without overriding any potential action's
icon:

    contentItem: IconLabel {
        icon: control.icon
        defaultIconColor: "tomato
    }

If icon.color was explicitly set, it is used instead.

This does mean that overriding the contentItem will result in the
icon color being lost, but that is already the case for style
customizations in general.

All controls that set icon.color and using IconLabel to display their
icon are affected and are therefore adapted: Button, CheckDelegate,
DelayButton, ItemDelegate, MenuBarItem, MenuItem, RadioDelegate,
RoundButton, SwipeDelegate, SwitchDelegate, TabButton, ToolButton.

Add FlowPane to the shared folder of the baseline test and use it in
the tests that we touch. This fixes two issues:
- Using Pane as the background fixes text being invisible for some
styles.
- Using Flow provides the most space-efficient layouting of items now
that we have more than would fit in the old layout.

Task-number: QTBUG-87459
Change-Id: I455ce7202a46b7cfa7545650574e48ad72796675
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>