Account for icon color being set to transparent
In QQuickAbstractButtonPrivate::updateEffectiveIcon we need to
resolve the icon's color so that QQuickIconLabel knows which of
defaultIconColor and icon.color should win.

In both QQuickAbstractButtonPrivate::updateEffectiveIcon and
QQuickAction::setIcon, we need to unconditionally set the icon
and be more strict about when we return early.

This fixes the case where setting "transparent" as the icon color
wouldn't result in the original image's colors being used, but
would instead result in the style's default icon color:

    Button {
        text: "Original icon color"
        icon.source: "heart.svg"
        icon.color: "transparent"
    }

    Button {
        text: "Original icon color (action)"
        action: Action {
            icon.source: "heart.svg"
            icon.color: "transparent"
        }
    }

Task-number: QTBUG-87459
Change-Id: Ic490e3b37a8174b8c034b84a1ad551a78088e44d
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>