ProgressBar: ensure the groove resepect the width of the control
As it stood, the progress bar groove would not be resized
according to the size of the control, since it lacked at
width binding.

This patch will make sure to resize the groove correctly. It
also fixes a binding loop warning related to the use of Loaders
in combination with childrenRect in the same go.

Fixes: QTBUG-141182
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I1aa87db339b04686ca521b486e4c7a05683a7679
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>