Slider/RangeSlider: ensure empty groove has the correct look
With Liquid Glass enabled, the groove of the Slider and
RangeSlider had a border which made the groove look taller
than the track/progress (which had no border). But looking
more carefully at the native NSSlider in Xcode, it shows
that the Slider groove (and track) should no longer have
any border at all (which is different from how it looks
in macOS 15).

This patch will therefore draw the track without a border when
running with Liquid Glass. Using nested rectangles in order to
draw the slider in the macOS 15 style was also unnecessary, as
we can achive the same look by using a single Rectangle.

Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I9d03a212f39857a448d103372636e9a12a7cbd3f
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>