qquickstyleconstants: notify that native colors may have changed after a theme change
When macOS changes appearance, we need to inform the
style that the system colors have changed. This patch
will therefore listen for changes to the colorScheme
and do just that.

Since the colors we return should match the effective
appearance, we need to wrap the NSColor calls inside
an NSApp.effectiveAppearance block.

Pick-to: 6.11 6.10 6.8 6.5
Change-Id: I282280909927136b85cd8f46f25892f825c8cb02
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>