SpinBox, macOS: make the stepper smaller
This is a 'cherry-pick' of
qtbase:b1a7a9d4f437c905df8368742b4189f1e8cea311

On macOS 26 this thing became rather large, not fitting into the
simple clipping rectangle of a spinbox (in a native app such a stepper is
taller than its controlled input, e.g. date/time edit). In Qt, with clipping
applied, this looks not so nice; but by always setting the cell's control size
to 'mini', we can have a properly looking spinbox, similar to what it was before.

Task-number: QTBUG-140134
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Ib157dc2f483a4fb10a4d99a68798367e0cd99a7a
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>