QQmlProperty: Split write handling for lists
The code working with lists backed by QQmlListProperty is quite
different than the one used to handle other types of lists. Separate the
two methods, which makes the function smaller, and which will allow more
focused unit testing later on.
Moreover, separate the code dealing with list properties in a part
dealing with the conversion logic, and a part which takes care of the
signal handling.

Task-number: QTBUG-140858
Pick-to: 6.10 6.8
Change-Id: Ie05cfca99fd0bdc668d320e1587abb68784ee1c2
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>