cmake: unify import paths for qmlls, qmllint, qmlimportscanner
Use _qt_internal_collect_qml_import_paths() to collect import paths for
qmlls, qmllint and qmlimportscanner.
_qt_internal_collect_qml_import_paths was previously only used to
collect import paths for qmlimportscanner.

Adapt _qt_internal_collect_qml_import_paths to add the special qmllint
import path, and remove the duplicate code for qmllint and qmlls's
import path.

Fixes QTBUG-141242 where the path of the QML module is not in the import
path of qmlls when the QML module path is a subfolder of the build
folder.

Pick-to: 6.10
Fixes: QTBUG-141242
Fixes: QTBUG-141555
Change-Id: I9fa2cc091ca08be6df9da5d6cf4794c08d3dd0ce
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>