Tooling: Add test for integrating event replay with qmlpreview
This makes sure that the relevant debug services can coexist. In order
for this to actually be true, we need to initialize the type loader
before we initialize the various services. Otherwise the profiler
service will initialize the type loader which will query the preview
service for some directories, leading to a deadlock.

Task-number: QTBUG-141569
Change-Id: Ia0873802b94b885654c693bdb9a96100cad492bf
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>