vectorimage: Fix transforms with constant animations
If an item has a "constant animation", i.e. an animation which
set the transform, but which is the same for its full duration,
we skip generating animations. However, these animations may
be set to replace previous transforms (the default transform
for instance). We add an optional branch which does this
programmatically in the case where we do not generate the
animations themselves.

This change also defines animations that start at an offset as
non-constant, since these should not be applied immediately
and therefore need a PauseAnimation.

Change-Id: I270b89c31093ac6e882a56f321f3fef487939038
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>