Fix flaky tst_qquickpathview::mouseDrag on Windows 11
Test is failing when the movingSpy.Size is checked before the dragging
had time to actually be completed.

Slightly increase the timeout period on QTRY_COMPARE when verifying
movingSpy.Size.

Previously QTRY_COMPARE would be trying for 1000 ms, increase the
timeout to 2000 ms to ensure that the pathview has finished being
dragged.

Change-Id: Ie28225067231dd523ef8136246c2796a2fe54cfa
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>