QQuickAttachedPropertyPropagator: fix type confusion
QQuickAttachedPropertyPropagator was trying to find a QQmlType for the
meta-object of an attached type. That would however only work in the
case of the attaching and attachment type being the same.
Add a helper function which always finds a possible QQmlType for a
meta-object. Note that it is possible to reuse the attachment type for
different attaching types, and we only return the first match. As an
optimization to not walk through all types, we also always return the
QQmlType of the passed-in type in case of self-attachments.

Moreover, because this lookup can still be rather expensive, we now
cache the result in QQuickAttachedPropertyPropagator.

Change-Id: Iecdde134ed5ca9420172be07f1c6ffeac827fc21
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>