baseline test: fix unstable grabbing of some animated scenes
For some animated scenes involving Shapes it has proven impossible on
some platforms/beckends to get a consistent grabbing of the final
stable state using the scengrabber's normal algorithm. Work around
this issue by letting the grabber check for an optional toplevel
property, thus allowing the scene itself to signal when it has reached
a steady state and is ready for grabbing.

Pick-to: 6.11 6.8
Change-Id: Iadfc03d23120dcb31b0e9e8ea1b5d2dee3f8073b
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>