QtQml: Don't delete the Qt object when clearing singletons
We need the object not only as singleton but also as member of the
global object. If there is a QJSEngine, parent it to the engine and make
it explicitly C++-owned. This way clearSingletons() won't delete it. If
there is no QJSEngine, there can't be any clearSingletons(). Therefore
we're safe in that case. The only known example for an ExecutionEngine
without a QJSEngine is the WorkerScript case, but since ExecutionEngine
is exported, it could be instantiated on its own in other Qt modules.

Amends commit 9dc7a22b212c18215942b9a4bfa17bd16dd5151b.

Pick-to: 6.10 6.8
Task-number: QTBUG-95788
Change-Id: I5a0aa35fadd552af208538ab2715efaff77f8f4e
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>