QmlModels: Clarify ownership of submodels stored in QVariant
Holding them in a local pointer makes it look like they leak. In actual
fact they are "owned" by the QVariant and assigned to a property later.

Also, check for model earlier. Apparently it can be nullptr if we're
dealing with a translation binding. In that case we don't hit the code
path that evaluates script bindings and therefore won't crash or leak.
However, the return statement after creating the QVariant is confusing
when trying to understand the ownership structure.

Coverity-Id: 897095
Coverity-Id: 897097
Change-Id: I770632d7e8aba7803ffe36363a6f9665873e3e61
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>