QmlModels: Make drain() safe against being called recursively
The callback can do pretty much anything, including another recursive
call to drain(). We need to make sure we're done with the container and
its iterators before we let that happen.

In the common case of maxPoolTime == 0, this is really easy. Simply swap
the container into a temporary and drain that. The case of maxPoolTime
!= 0 is a bit more complicated, but the algorithms library gives us a
way to do it with std::stable_partition and std::move that should be
good enough.

Fixes: QTBUG-102811
Pick-to: 6.10 6.8 6.5
Change-Id: Id67a9b9c2a7279536e5c2a315b139fcd8024e243
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>