qmllint: Fix formatting for --list-plugins options
The plugins details can extend outside of their respective column if
they are too wide. This also offsets the remaining columns on that line,
breaking the formatting of the table.

Instead of simply allocating a fixed space for each element, compute the
required width of each column and use it to format the table.

Amends 7c9276d38b311042235a476e7421b64549b43413

Pick-to: 6.10 6.8 6.5
Change-Id: I67a5484eda38cc96bb5c66ba7283b87c7f1bb3a1
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>