Make the rate at which to update hover and cursor configurable
Profiling showed that in a mostly idle application, if the mouse cursor
was hovering and something (such as an animation) caused some item to
become dirty periodically, we would spend a lot of CPU time sending
fake hover events and checking whether any animated item or HoverHandler
requires the mouse cursor to change (for example by moving under or
away from the static mouse cursor position). We needed to do it because
we got bug reports about hover state of animated items moving under the
cursor; but we don't want to spend so much time on it. It's unlikely to
be noticed in most applications, but we have been spending time on it in
every application anyway, just because of a few complaining users.
Delivering so many hover events (beginning with
6f84a09dfbd15aac023580cf06e7b8c24f3b524c) was inefficient enough, and
then since cd7c5f94a0ed64c52d31998957134d099313932a we do yet another
traversal to check for cursor changes.

The human threshold for visual perception of change or movement is about
1/10 second: so now we update the cursor at a minimum interval of 100 ms
by default (but it's still frame-synchronous). In other words, if the
screen refresh is 60 FPS, we do it 1/6 as often as we did before, which
reduces CPU usage. It's also configurable via an environment variable,
in case that's for some reason not suitable for someone, or someone
wants to turn off this feature completely.

If a non-periodic change causes an item to become dirty, a frame will be
rendered: and if at that time, the elapsed time is not sufficient,
frame-synchronous hover event would not be delivered. We start the
new frameSynchronousDelayTimer to call flushFrameSynchronousEvents()
later.

frameSynchronousHoverInterval replaces frameSynchronousHoverEnabled:
set it to -1 to disable frame-synchronous hover completely.

Amends cd7c5f94a0ed64c52d31998957134d099313932a

[ChangeLog][QtQuick][Hover/Cursor] You can now configure the minimum time
interval between updating mouse hover and cursor when items animate
under it, by setting QT_QUICK_FRAME_SYNCHRONOUS_HOVER_INTERVAL to a
number of milliseconds. By default we use 100 milliseconds, to reduce
CPU usage, rather than doing it before every frame rendered, as we did
in previous Qt versions. It can also be set to -1 to send hover events
and update the cursor only when the mouse moves, not when something
moves under a stationary cursor.

Pick-to: 6.11
Task-number: QTBUG-53987
Task-number: QTBUG-54019
Task-number: QTBUG-90457
Task-number: QTBUG-136976
Task-number: QTBUG-140340
Change-Id: Iec81920e220232d557f607b32e6d04dc7995c929
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>