SearchField: Prevent clipped search icon on macOS 26(Tahoe)
macOS 26 (Tahoe) changed NSSearchFieldCell: the magnifying glass icon is
taller. Drawing it at the button's full rect causes the top to be clipped
by the bezel, though the clear icon remains fine. To avoid this, we
render the search icon at a fixed, smaller size to prevent upscaling
and eliminate clipping after 26.0.

Fixes: QTBUG-141776
Pick-to: 6.10
Change-Id: Ifb5f4476253e55e700b8b89e864afdbb2ceedd8f
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>