QmlPreview: Allow setting animation speed
This is the same functionality we already have in the inspector service.
However, we need it in the preview, too. For such a simple thing you
shouldn't need to instantiate an extra debug service.

The animation speed method in the preview service also doesn't invert
the speed modifier. So, for double animation speed you need to pass 2,
not 0.5.

Fix the test to a, check the animation speed only on new output, rather
than everything the application has printed before and b, actually check
whether the number is in the intended range.

Task-number: QTBUG-141569
Change-Id: Ie7f7b043fb295e7c5da333ab4501059f07f2e3cd
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>