QtQml: Reset JS libraries and ES modules on clearComponentCache()
We can't drop the respective compilation units without re-opening
QTBUG-128638 and QTBUG-128782. However, we can empty them out and force
them to be re-evaluated. Since you can store arbitrary data in those
libraries, this is important for a proper clearing of the component
cache. You may hold on to a QObject that in turn holds on to other
compilation units.

Change-Id: Ie6bd3878c0d45086a7efade76ad5696bf8428acb
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>