tst_qqmllanguage: Don't pollute the source directory
The i18n tests that QML files are always UTF-8. To do this is creates a
QML file with a name that only makes sense if it is UTF-8 encoded and
tries to load it from another file.

This file is generated next to the other test files for the i18n tests
in the source directory. The files is removed at the end of the test by
some cleanup code but if the execution is interupted during the test,
like during a compile-debug loop, the file is left behind and pollutes
the source directory.

Rework the test to utilize a temporary directory. This way any files
left behind do not pollute the source.

The check was initially introduced by
176698a91d3b2256a093785967c6706328981af4

Later 1ec1a863471445132c1d29500d6ac720663887ae adapted it to fix it on
Android. The .qrc and .pro files it introduced or edited were later
removed by 338e892a18b73c912881e6e7f93ade0dbaa1eb8b and
78ab4b8d8b75fed240cded1a3f182048e56a8823 respectively.

Amends 176698a91d3b2256a093785967c6706328981af4

Change-Id: I8175e1e335a06d1389f00c4fd5905ba69be87f31
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>