gc: Fix sweep logic
When sweeping chunks, we might end up allocating additional chunks in a
custom destroy handler. This could cause the insertion of a new chunk
into the chunk vetcor, in the worst case invalidating the iterators (in
the slightly less worse case, we still have a chunk on which we don't
call sweep, leading to meta-data corruption later on).

Note that while the number of chunks can grow, it can't decrease.
Therefore, we can fix the problem by replacing the iterator based
partition with a size based one, which has additional checks in case the
size has changed during the iteration.
As a small optimization, wer run the loops without always checking for a
size change first, and only afterwardts check for it (as allocating
destroy handler aren't that common). Only afterwards we check if the
size has changed, and rerun the loop on the remaining vector.

Pick-to: 6.10 6.8 6.5
Initial-patch-by: Rafal Chomentowski <rafal.chomentowski@ge.com>
Fixes: QTBUG-141963
Change-Id: If9429a4a9030ec64952426ac35c38d3ae250e18d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>