Window: specify correct property type
Given that the property was documented to only take a screen from
Application's screen list, we can arguably do this change, especially
that any usage with a non-QQuickScreenInfo would just end up setting a
null-pointer, indicating some bug.
Any code reading the property is fine anyway, as a QQuickScreenInfo is
still a QObject.

This also uncovers that the screengrabber base line test did not handle
this correctly (it passed a QScreen to setScreen, which doesn't work).
This would now fail to compile. Fix it by passing the actually expected
object along.

Amends d6dfbe8fd5d09388fc624d85c4aefa0269c20db9

[ChangeLog][Important Behavior Change][QtQuick][Window] Window and
ApplicationWindow now enforce that the type of the property passed to
the screen property is ScreenInfo. It used to accept any QObject, but
would silently convert it to null if the type was not correct.

Change-Id: I512c24c761ef5e6e9b7f241d222c98a34c6fafa4
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>