Dialog: Don't give buttonBox focus when empty or buttons are disabled
Focus behavior was changed in 4d1753b95f0343a68f80a1b482b8ab9a0397fecb
to give the buttonBox focus when a dialog is opened.
This will fundamentally conflict with user projects that manually set
the focus property.

Further discussion will be needed to arrive on a good solution for when
the focus property is manually set. For now, we can at least make sure
that the new focus behavior is limited to cases where the buttonBox
actually contains buttons (aka, not empty), with at least one of them
being enabled.

Fixes: QTBUG-141998
Change-Id: I704e6a707f5b579e5f4de7b6197abc2fc5dfdf3e
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>