VectorImage: Implement support for gaussian blur
This introduces support for the gaussian blur filter using
the MultiEffect as an approximation. The approximation is
not going to look perfect, but will perform better than
doing an actual gaussian blur on the contents. Since the
objective here is runtime performance, this is a reasonable
trade-off.

In order to make this work correctly with masks, we also
refactor the mask code and now use ShaderEffectSources
instead of layers in order to have full control.

Task-number: QTBUG-121541
Change-Id: Idb1574a2f4d03311ab787fb4a6476b6c604c88b2
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>