qmlls: qInfo() when receiving EOF
Add a test to avoid QTBUG-142025 from regressing, and log a message with
qInfo() when receiving EOF to simplify debugging qmlls in the future.
This should help to tell if qmlls stopped on start because of
QTBUG-142025 or other reasons (invalid signature, or the qmlls client
trying to spawn qmlls from a filepath that does not exist for example).

The actual fix of QTBUG-142025 is in qtbase and tracked via QTBUG-142041.

Fixes: QTBUG-142025
Pick-to: 6.10
Change-Id: I9cd215ac02e23a1d833755eda058d72622de7318
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>