QmlModels: Add RAII wrapper for "script" references
Script references to QQmlDelegateModelItems were references held by
either the created delegate object, the incubation task, or the heap
object that holds the QQmlDelegateModelItem. However, for the object and
the incubation task we only incremented the script reference counter if
they were not populated by QQmlTableInstanceModel. Indeed it makes very
little sense to hold a script reference for those in general. The
incubation task is explicitly checked on isReferenced() and for the
object we have a separate reference count.

Encapsulate the reference counting and drop the unnecessary references.
Check the object in addition to isReferenced() where it matters.

Task-number: QTBUG-141963
Change-Id: Id0fcb7d36afcfc131f52584fc9bcab8989ff7fc1
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>