Compiler: Stop std::move-ing the changed register in certain cases
This patch removes an optimization where we std::move the changed
register into its target. A copy is performed instead. This resolves
the more urgent issue of miscompilations.

QTBUG-141920 has popped up and was either exposed or introduced by
b303e0624d2ea2ab1c124961e7510a64d0ca1412.

That change performs optimizations that avoid intermediary copies of
values between registers. It was found that, in the process, the mapping
from RegisterContent to variable name stored in m_resiterVariables was
changed. One RegisterContent no longer uniquely maps to a variable name
because of the direct moving of values to their destination. The
registers alias the same value. This could cause issues and might need
to be addressed.

The original std::move was only added after 6.5 by
3193911b02424dd0365e03526a4c12ed7888b7ca. Cherry-pick to the later
versions.

Fixes: QTBUG-141920
Pick-to: 6.10 6.8
Change-Id: Ie6dad1a809ac5ee08de70971ec92d3c97fc50080
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>