CMake: QML: put test targets under the special case section
Arguably targets relevant to tests should be built once all the "source"
targets have been built, since they might depend on them.
For example, in the near future there is a need to include QmlDomPrivate
target to the QuickTestUtils, which is defined only later.

Moreover it's in general a good practice to have a clear separation between
"source" and "test" targets.

Change-Id: I2dd1b41ee4df2227733bab73f983e001daad305c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>