Fix several flaky tst_QQuickStyle on different platforms
Fix the flaky tests tst_QQuickStyle::commandLineArgument,
configurationFile, environmentVariables. The tests were picking up the
previously selected style, causing them to be unable to change and to
fail the next tests. When re-running a test no style is selected yet,
meaning the desired test can be selected succesfully.

Change the order to have lookup as the last time as it will make
other tests fail the first time. The lookup test is still failing the
first time it is run due to the wrong font size.

Reset the style inside of cleanup()

Change-Id: Ic974316c856affbc5f3fe805aabd27588b021c4c
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>