VectorImage: Add support for more filters
This adds support for the feColorMatrix, feOffset and
feFlood filters.

We also need to change how we calculate the bounds of
post-processed items to use the internal bounds. Basically,
we need the bounds of the children of the item + the
item itself before the node's transform is applied. This
is because the bounding box is axis aligned, so we cannot
get the original rectangle back just by applying the
inverse transform in the case of e.g. rotations.

Task-number: QTBUG-121541
Change-Id: I7a8863bef4397df4043c854256dca312b91f888a
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>