qmllint: Remove LoggerCategory isDefault but add isEssential
We want the user to have more control over which categories are enabled
and at which severity when linting. The default flag prevents the user
from doing that. Get rid of it and allow the user to change the severity
of categories both up and down.

In order for qmllint to function properly, some categegories are still
required. Mark them as essential. The only such category at the moment
is qmlSyntax. Without it, other categories might fail and silently
misbehave.

Task-number: QTBUG-141233
Change-Id: I458c6b728e3933cd0dcd855052e137f667a75c94
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>