QQStyleKitControl: add support for Instance Variations
This patch will add support for Instance Variations. Instance
Variations allows you to define one or more style variations in
a StyleKit style that can be activated from the application
using an attached 'StyleKitControl.variations' object.
Instance Variations will affects all descendant StyleKitReaders
of the item that contains the attached object.

For example, if you set "StyleKitControl.variations: ['mini']"
on a GroupBox, all controls inside that GroupBox will
be affected with the variation named "mini" in the style
(if any).

Inside a variation, you specify which controls should
receive alternative styling when the variation is applied.
Any properties defined in a Variation override those
set in the Style or Theme.

In order to support Instance Variations, this patch will also
refactor how we implement Type Variations, which are variations
that applies to _all_ controls of a specific type, rather
than to individual instances.

Change-Id: I6486979281997e69b65da0ed4866b264c91c592f
Reviewed-by: Doris Verria <doris.verria@qt.io>