Change order tst_palette::comboBoxPopup to ensure theme is not fixed
Tst_palette::comboBoxPopupWithThemDefault fails on its first run
because tst_palette:comboBoxPopup sets a theme beforehand. It always
passes on its second run, as no theme as been fixed yet.

Change the order of the two tests ensuring that
comboBoxPopupWithThemDefault is passing on its first run.

This is a workaround to prevent the theme being fixed before the test.

Change-Id: I902f6b0cff9f35305cb5534d1318395677fa2748
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>