QJSEngine: fix -Wnrvo in coerceValue() and fromVariant()
... as a preparation of enabling the flag in headerscheck.

The coerceValue() functions was actually flagged by GCC in a
headerscheck including -Wnrvo, probably because it was instantiated
for some type in inline code. The fromVariant() function was not
flagged automatically, but found by manual inspection.

Because the structure of code, with the pre-existing extra scope,
lends itself naturally to it, wrap the code that doesn't already
return rvalues in an IILE. Outside the lambda, the return is now
RVO'ed, and inside, NRVO'ed.

Manual inspection also found the qjsvalue_cast() overloaded functions,
but I don't readily see a fix for them, so deferring. GCC doesn't warn
for function templates unless instantiated, so for the purposes of
headerscheck, we should be good with the status quo.

Amends 7b29ed6f64a300d4d714371c2872fcba344beebd (coerceValue()) and
d0f4e0c037cf61eb5bb559755ee7c9ce6cf6b7dc (fromVariant()).

Pick-to: 6.10 6.8 6.5
Task-number: QTBUG-142146
Change-Id: I2aa0ecb5a1cdc975c2ab8dffb9fe333bbeb7bc0f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>