Qml Profiling: Handle the case of profiling without any features
We cannot encode the "running" state of a profiling adapter into the
features it's recording. It may be started with no actual features
enabled and then we still need to stop it eventually.

In order for the QmlProfiler test to work without the DebugMessages
feature we still need to create the DebugMessages client, but ignore its
output. Otherwise the logic that checks "unrelated" clients gets
confused.

Change-Id: I52a3c2bc55ea36454ab9c8c7f25fe833c7f14a43
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>