a11y: Make QAccessibleQuickItem report the application's locale
QAccessibleQuickItem now also implements QAccessibleAttributesInterface
and returns the QLocale of QQuickApplicationWindow if it is parented
to such a window. Otherwise an empty QVariant is returned.

This is helpful for Screen Readers to read texts in the application's
language in case it differs from the OS language.

For QQuickControls this is already implemented as they also feature
a separate `locale` property.

Task-number: QTBUG-141701
Pick-to: 6.11
Change-Id: Ie9630d5f02964d9dcd3da72fe0c5f20a3376276e
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>