qmlls: use 1 job for background builds
Control the numbers of jobs to be used in Qmlls in background builds via
the "-j"  argument of CMake. Set the default to 1: a background build
should not eat up all the CPU/RAM available of the users computer.

[ChangeLog][qmlls] The number of CMake jobs can be set with a command line
option, an environment variables, and a .ini setting option. The default
was changed from using all available cores to 1. You can restore the
previous behavior by passing "max" as number of CMake jobs.

Task-number: QTBUG-142352
Change-Id: Icc8573906632272773c0a2812febbf06aa6b3079
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>