VectorImage: Fix opacity on individual shapes
When shapes are combined into a single Shape object, the opacity information
is lost. Similarly to how we handle non-default transforms on the shapes, we
make sure the shapes with opacities are never combined with others.

In principle, we could combine shapes with the same opacity, but for simplicity
we just split them up at least until we see a strong need.

Pick-to: 6.10
Change-Id: I2c8a444f9b38fc8cf66728d80b1a724f24469630
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>